Fix #37 Datamanager setChannel
Implement FrameworkInfo/Profiling

After deleting an object in DataManager::setChannel the following new
created an object at the exact same position. This prevented a segfault
but was at the same time undefined behavior. When executing LMS under
valgrind this undefined behavior lead to an invalid read in the data channel.

I fixed this issue by using copy assignment instead of deleting the
old PointerWrapper and creating a new one.